+++ branches/next_ghost/libs/bgraph.h
@@ -74,7 +74,7 @@ public:
 	const Glyph &glyph(unsigned idx) const;
 	unsigned textWidth(const char *text) const;
 	unsigned textHeight(const char *text) const;
-	unsigned memsize(void) const;
+	size_t memsize(void) const;
 };
 
 class SoftRenderer : public Texture {
@@ -131,7 +131,7 @@ public:
 	virtual void hideMouse(void) { }
 	virtual void moveMouse(int x, int y) { }
 
-	unsigned memsize(void) const { return Texture::memsize() + sizeof(*this) - sizeof(Texture); }
+	size_t memsize(void) const { return Texture::memsize() + sizeof(*this) - sizeof(Texture); }
 };
 
 class FadeRenderer : public TextureFade {
@@ -153,7 +153,7 @@ public:
 
 	const TexturePal &operator[](unsigned idx) const;
 	unsigned size(void) const { return _count; }
-	unsigned memsize(void) const;
+	size_t memsize(void) const;
 };
 
 // TODO: write a complete video backend instead
